The projetc that will represent the fixed income controller.
This project will manage the existing entries, allow the addition of newones
and handle the events that can ocurs with the indexers. The projections stay
in FixedIncomeData.

I still don't know witch will be the persistence but for now I'll use a text
file and maybe a sqlite interface afterwards.